\name{itamtplcost}
\alias{itamtplcost}
\title{Italian MTPL cost}
\docType{data}

\description{
 This dataset contains large losses (in excess of 500 Keuro) of an Italian Motor-TPL company since 1997.
}
\usage{
data(itamtplcost)
}
\format{ 
\code{itamtplcost} is a data frame of 2 columns and 457 rows:
\describe{
	\item{\code{Date}}{Date of loss (accident date).}
	\item{\code{UltimateCost}}{Ultimate cost trended to 2013 and 
		developed to ultimate losses.}
}
}

\source{ 
Unknown private insurer.
 }
 
\examples{

# (1) load of data
#
data(itamtplcost)


}
\keyword{datasets}