\name{swtriangles}
\alias{swtri1auto}
\alias{swtriangles}

\title{Switzerland general liability triangles}
\docType{data}


\description{

\code{swtri1auto} is a named list of two triangles : the incurred (cumulative) amounts
and the paid (cumulative) amounts.

}
\usage{
data(swtri1auto)

}
\format{ 
\code{swtriangles} is a named list of two matrices, respectively for incurred and paid amounts.
}


\references{ 

Dahms, R. (2008),
\emph{A Loss Reserving Method for Incomplete Claim Data}, 
Bulletin of the Swiss Association of Actuaries, pp. 127-148.

Dahms, R., Merz, M., Wuethrich, M.V. (2009), 
\emph{Claims development result for combined
claims incurred and claims paid data}. Bulletin Francais d'Actuariat 9 (18), 5-39.

Merz, M., and M. V. Wuethrich (2010),
\emph{Paid-Incurred Chain Claims
Reserving Method}, Insurance: Mathematics and Economics
46, 2010, pp. 568-579.

Merz, M., and M. V. Wuethrich (2013), 
\emph{Estimation of Tail Development Factors in the Paid-Incurred Chain
Reserving Method}, Variance 71, pp. 61-73.


 }
 
\examples{

# (1) load of data
#
data(swtri1auto)




}
\keyword{datasets}