\name{ushustormloss4980}
\alias{ushustormloss4980}
\title{Normalized Hurricane Damages in US between 1949 and 1980}
\docType{data}


\description{

	Normalized Hurricane Damages in the United States due to single hurricanes. 
	They applied to the period from 1949 and 1980 and are adjusted for inflation.
	Originally, the dataset was compiled by the American Insurance Association
	and is also reported in Beirlant, Teugels and Vynckier (1996).
	

}
\usage{

data(ushustormloss4980)

}
\format{ 
\code{ushustormloss4980} is a data frame of 7 columns and 207 rows:
\describe{
	\item{\code{NormLoss80}}{Normalized damages (million of 1980 USD).}

}

}

 
\references{ 
Dataset used in Beirlant, Teugels and Vynckier (1996),
\emph{Practical Analysis of Extreme Values},
Leuven University Press. 

 }
 

\examples{

# (1) load of data
#
data(ushustormloss4980)


}
\keyword{datasets}