\name{FedYieldCurve}
\alias{FedYieldCurve}
\docType{data}
\title{Federal Reserve interest rates }
\description{
The data-set contains the interest rates of the Federal Reserve, from January 1982 to December 2012.
The interest rates are Market yield on U.S. Treasury securities constant maturity (CMT)
(more information on the Treasury yield curve can be found at the following website
\url{http://www.treasury.gov/resource-center/data-chart-center/interest-rates/Pages/yieldmethod.aspx})
at different maturities (3 months, 6 months, 1 year, 2 years, 3 years, 5 years, 7 years and 10 years),
quoted on investment basis and have been gathered with monthly frequency.
}
\usage{data(FedYieldCurve)}
\format{
  An object with class attributes \code{xts}.
}
\source{
\acronym{FED}: \url{http://www.federalreserve.gov/datadownload/Build.aspx?rel=H15}.
}
\keyword{datasets}