\name{linearmodelfactor}
\alias{linearmodelfactor}

\docType{data}
\title{ A simulated with linear model factor }

\description{ A simulated with linear model factor }


\usage{
data(linearmodelfactor)

}

\format{ 

The variables for \code{linearmodelfactor} are
\describe{
    \item{\code{X}}{A numeric.}
    \item{\code{Y}}{A numeric.}
    \item{\code{Z}}{A factor.}
	}


}


\examples{

# (1) load of data
#
data(linearmodelfactor)


head(linearmodelfactor)




}

\keyword{datasets}
