\name{usautotriangles}
\alias{usautotri9504}
\alias{usautotriangles}

\title{US Automobile triangles}
\docType{data}


\description{

\code{usautotri9504} comes from Wacek (2007) and represent 
industry aggregates for private passenger
auto liability/medical coverages.
This dataset contains cumulative payments between 1995 
and 2004 in millions of dollars.
Amounts are based on insurance company annual statements from
Schedule P (Part 3B). The elements of the triangle represent cumulative
net payments, including defense and cost containment expenses.


}
\usage{
data(usautotri9504)


}
\format{ 
\code{usautotri9504} is a matrix
containing insurance triangles.

}

\source{ 
\href{http://instruction.bus.wisc.edu/jfrees/jfreesbooks/}{FreesBook-RMAFA}

 }

\references{ 


Frees (2010), 
\emph{Regression modelling with actuarial and financial applications},
Cambridge University Press.

Wacek, M.G. (2007). 
\emph{The path of the ultimate loss ratio estimate},
 Variance 1, no. 2, 173-92.


 }
 
\examples{

# (1) load of data
#
data(usautotri9504)




}
\keyword{datasets}